Various improvements
Created a stash model for the test stash/retrieve images actions.
(This used to be the old image model).
Improved stash action when there were no more images to stash.
Improved the homepage with links to different pages.
Improved the stats page, including a properly centered container and box shadow.